backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pdfplumber
groq
langchain
langchain-groq
langgraph
faiss-cpu
numpy
sentence-transformers
python-dotenv
pydantic
